fix(docs): volume mount path
Using `/paper/install/dir` as a static volume mount path caused
permission issues for the user `minecraft` within the container.
So we let docker create the volume and handle permissions.

For this static mount path to work, we would first need to create a
matching user (uid must be the same) on the host system and let that uid
own the volume directory.